refact(editor): clipboard config should be in foundation (#12038)
