+++ blocksuite/affine/blocks/root/src/clipboard/readonly-clipboard.ts
@@ -1,12 +1,6 @@
 import {
-  AttachmentAdapter,
-  ClipboardAdapter,
   copyMiddleware,
   defaultImageProxyMiddleware,
-  HtmlAdapter,
-  ImageAdapter,
-  MixTextAdapter,
-  NotionTextAdapter,
   titleMiddleware,
 } from '@blocksuite/affine-shared/adapters';
 import {
@@ -15,67 +9,7 @@ import {
   getSelectedModelsCommand,
 } from '@blocksuite/affine-shared/commands';
 import { DisposableGroup } from '@blocksuite/global/disposable';
-import {
-  ClipboardAdapterConfigExtension,
-  LifeCycleWatcher,
-  type UIEventHandler,
-} from '@blocksuite/std';
-import type { ExtensionType } from '@blocksuite/store';
-
-const SnapshotClipboardConfig = ClipboardAdapterConfigExtension({
-  mimeType: ClipboardAdapter.MIME,
-  adapter: ClipboardAdapter,
-  priority: 100,
-});
-
-const NotionClipboardConfig = ClipboardAdapterConfigExtension({
-  mimeType: 'text/_notion-text-production',
-  adapter: NotionTextAdapter,
-  priority: 95,
-});
-
-const HtmlClipboardConfig = ClipboardAdapterConfigExtension({
-  mimeType: 'text/html',
-  adapter: HtmlAdapter,
-  priority: 90,
-});
-
-const imageClipboardConfigs = [
-  'image/apng',
-  'image/avif',
-  'image/gif',
-  'image/jpeg',
-  'image/png',
-  'image/svg+xml',
-  'image/webp',
-].map(mimeType => {
-  return ClipboardAdapterConfigExtension({
-    mimeType,
-    adapter: ImageAdapter,
-    priority: 80,
-  });
-});
-
-const PlainTextClipboardConfig = ClipboardAdapterConfigExtension({
-  mimeType: 'text/plain',
-  adapter: MixTextAdapter,
-  priority: 70,
-});
-
-const AttachmentClipboardConfig = ClipboardAdapterConfigExtension({
-  mimeType: '*/*',
-  adapter: AttachmentAdapter,
-  priority: 60,
-});
-
-export const clipboardConfigs: ExtensionType[] = [
-  SnapshotClipboardConfig,
-  NotionClipboardConfig,
-  HtmlClipboardConfig,
-  ...imageClipboardConfigs,
-  PlainTextClipboardConfig,
-  AttachmentClipboardConfig,
-];
+import { LifeCycleWatcher, type UIEventHandler } from '@blocksuite/std';
 
 /**
  * ReadOnlyClipboard is a class that provides a read-only clipboard for the root block.
