+++ blocksuite/affine/blocks/root/src/common-specs/index.ts
@@ -4,14 +4,12 @@ import { BlockFlavourIdentifier, FlavourExtension } from '@blocksuite/std';
 import type { ExtensionType } from '@blocksuite/store';
 
 import { RootBlockAdapterExtensions } from '../adapters/extension';
-import { clipboardConfigs } from '../clipboard';
 import { builtinToolbarConfig } from '../configs/toolbar';
 import { fallbackKeymap } from '../keyboard/keymap';
 
 export const CommonSpecs: ExtensionType[] = [
   FlavourExtension('affine:page'),
   ...RootBlockAdapterExtensions,
-  ...clipboardConfigs,
   fallbackKeymap,
 
   ToolbarModuleExtension({
